+++ packages/datatrak-web/src/components/index.ts
@@ -1,13 +1,7 @@
 // Exporting these first to avoid issues with circular references
 export * from './Icons';
 export { Modal } from './Modal';
-export {
-  PageContainer,
-  SafeArea,
-  SafeAreaColumn,
-  SafeAreaFooter,
-  SafeAreaHeader,
-} from './PageContainer';
+export { PageContainer } from './PageContainer';
 
 export { Autocomplete, QuestionAutocomplete } from './Autocomplete';
 export { Button } from './Button';
